requirements/requirements-all.txt
attmap>=0.12.5
logmuse>=0.2
pandas>=0.24.2
pyyaml
ubiquerg>=0.5.2
